+++ src/Application/Services/WebService/ZarinPal/IZarinPalWebService.cs
@@ -5,10 +5,10 @@ namespace DigitalWallet.Application.Services.WebService.ZarinPal
 {
     public interface IZarinpalWebService
     {
-        Task<(IRestResponse Response, ResultZarinPal<PaymentRequestResultZarinPal> Result)> PaymentRequestAsync(double amount,
+        Task<(IRestResponse Response, ResultZarinPal<PaymentRequestResultZarinPal> Result)> PaymentRequestAsync(decimal amount,
             string? description = null, string? mobile = null, string? email = null);
 
-        Task<(IRestResponse Response, ResultZarinPal<ZarinpalVerifyPaymentResponse> Result)> VerifyPaymentAsync(double amount,
+        Task<(IRestResponse Response, ResultZarinPal<ZarinpalVerifyPaymentResponse> Result)> VerifyPaymentAsync(decimal amount,
             string authority, CancellationToken cancellationToken = default);
     }
 }
