+++ src/Presentations/WebApi/Models/WalletDto.cs
@@ -13,7 +13,7 @@ public class IncreaseDto
 {
     [Required(ErrorMessage = "Wallet identity cannot be null.")]
     public Guid WalletId { get; set; }
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
     public string? Description { get; set; } = null;
 }
 
@@ -21,7 +21,7 @@ public class DecreaseDto
 {
     [Required(ErrorMessage = "Wallet identity cannot be null.")]
     public Guid WalletId { get; set; }
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
     public string? Description { get; set; } = null;
 }
 
@@ -30,7 +30,7 @@ public class DepositDto
     [Required(ErrorMessage = "Wallet identity cannot be null.")]
     public Guid WalletId { get; set; }
     [Range(100, 500000, ErrorMessage = "Amount must greater then {1} and smaller than 500,000.")]
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
     [Required(ErrorMessage = "Trace id cannot be null.")]
     public string? TraceId { get; set; }
     [Required(ErrorMessage = "Callback cannot be null.")]
